requirements.txt
absl-py==2.2.0
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
anyio==3.7.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
audioread==3.0.1
backcall==0.2.0
bert-score==0.3.13
bitsandbytes==0.44.1
black==24.8.0
cachetools==5.5.0
captum==0.7.0
catboost==1.2.7
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
chess==1.11.2
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.8
decorator==5.1.1
diffusers==0.33.1
dill==0.3.8
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.1.0
faiss-gpu==1.7.2
fastdtw==0.3.4
fastjsonschema==2.21.1
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.70.0
h11==0.14.0
h5py==3.11.0
hdbscan==0.8.40
hf-xet==1.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.31.1
identify==2.6.1
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.3
jiter==0.9.0
joblib==1.4.2
joypy==0.2.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kagglehub==0.2.9
keras==2.13.1
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
lightgbm==4.6.0
lightning-utilities==0.11.9
lion-pytorch==0.2.3
llvmlite==0.41.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mne==1.6.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
ninja==1.11.1.4
nltk==3.9.1
nodeenv==1.9.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
oauthlib==3.2.2
openai==1.68.2
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
peft==0.13.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
pooch==1.8.2
pre-commit==3.5.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.6
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.4
pyro-api==0.1.2
pyro-ppl==1.9.1
PySoundFile==0.9.0.post1
pytest==8.3.5
python-chess==1.999
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rouge==1.0.1
rpds-py==0.20.1
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.5
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentence-transformers==3.2.1
sentencepiece==0.2.0
sentry-sdk==2.24.1
setproctitle==1.3.5
shap==0.44.1
six==1.16.0
slicer==0.0.7
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.0
soxr==0.3.7
squarify==0.4.4
stack-data==0.6.3
statsmodels==0.14.1
stockfish==3.28.0
style==1.1.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tiktoken==0.7.0
tokenizers==0.20.3
tomli==2.2.1
torch==2.4.1
torchaudio==2.4.1
TorchCRF==1.1.0
torchmetrics==1.5.2
torchvision==0.16.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.3
triton==3.0.0
typing_extensions==4.13.0
tzdata==2024.2
umap==0.1.1
umap-learn==0.5.7
update==0.0.1
urllib3==2.2.3
virtualenv==20.31.1
wandb==0.19.8
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
xgboost==2.1.4
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2